Add force_bridge option permitting use of "down" adapters
There is a legitimate case for being able to use an adapter that is "down" for a bridge. 

For reasons of host isolation, it is perfectly acceptable to have an adapter physically connected to a network on a virtualbox host, but in the host OS have no configured IPv4, IPv6, nor any other protocol, so that the host OS remains unreachable on that particular physical network, perhaps for security reasons. 

This means the network adapter is specifically reserved for connecting virtual machines to a physical network via bridging. To the OS and VBoxManage, this adapter appears as "Down" (due to the lack of IP address), however, this host adapter is still very capable of being used to connect virtual machines to the physical network. The way this was before, the host had to have the adapter be "up" (have an IP address) in order for a vagrant virtualbox guest to even hope of using the adapter.

This change adds a "force_bridge" parameter to the network config that permits network adapters to be used by virtual machines for bridging while the host OS remains isolated. The force parameter ignores the interface status and assumes that if you have gone to all the trouble to force the use of the adapter, you must know what you really want to have happen.